refactor(callbacks): split register_callbacks — Alerts + close-out (Step 10d)
Final step of the ``register_callbacks`` decomposition (issue #87,
Step 10d). Builds on PRs #109 (Overview + US Grid) and #110 (Models +
Forecast). Moves the last cleanly tab-bounded callback block —
Risk / Alerts — into ``_callbacks_alerts.py``. Closes the issue.

## What moves in this PR

Three Risk tab callbacks (262 lines):

* ``update_risk_title`` — page title block.
* ``update_risk_insight`` — 1-sentence narrative summary.
* ``update_alerts_tab`` — 8-output payload (alert cards, stress
  score / label / breakdown, anomaly chart, temperature exceedance,
  historical event timeline, weather-context KPIs) with Redis fast
  path + v1 demo-data compute fallback.

All move to ``register_alerts_callbacks(app)`` in
``components/_callbacks_alerts.py``.

## Drive-by fixes

* **Renamed stale section headers in callbacks.py.** The Alerts
  block was labeled ``# ── 7. TAB 4: GENERATION & NET LOAD ──`` —
  another stale breadcrumb from an earlier refactor that deleted
  the Generation TAB 4 callbacks but left the comment. Renamed to
  ``# ── 5. RISK / ALERTS TAB ──``. The subsequent ``# ── 9. TAB 6:
  SCENARIO SIMULATOR ──`` actually marked the start of the fallback
  banner callback (cross-cutting, stays in callbacks.py) — renamed
  to ``# ── 6. FALLBACK BANNER ──``.
* **Re-added ``COLORS``, ``_empty_figure``, ``_layout``, ``REGION_NAMES``
  to the shared re-export block** with ``# noqa: F401`` markers.
  ruff auto-fix stripped them after the Alerts extraction removed
  their last internal consumer in callbacks.py, but tests still
  reach in via ``cb.X`` / patch them via ``components.callbacks.X``.

## Decomposition close-out

Updated ``docs/internal/CALLBACKS_DECOMPOSITION_PLAN.md`` from
forward-looking plan to close-out report. Documents:

* Final file inventory (callbacks.py: 1,003 lines; 9 per-tab modules
  totaling 7,164 lines)
* Per-PR delta table (#98–#111, 14 PRs total)
* Acceptance-criteria scorecard
* Five "surprises we hit" (stale section comments, patch-where-it's-
  used gotcha, ruff F401 shim wars, Forecast↔Backtest cross-dep
  factoring, orphaned Generation+Weather fast paths)
* Effort estimate vs actual (half-day → 31 hours across 14 PRs)
* What was intentionally not done

## Size impact

callbacks.py: 1,265 → 1,003 (–262 lines, –20.7%)
_callbacks_alerts.py: 260 → 564 (+304 lines)

**Total decomposition: 7,167 → 1,003 lines in callbacks.py (–86%).**

## What's left in callbacks.py

Cross-cutting callbacks only — data loading, persona switching,
URL state (bookmarks, smart defaults, cross-tab links), header
freshness, fallback banner, confidence badges, briefing mode.
These are foundational rather than tab-specific. Splitting them
into a ``register_shared_callbacks`` would add ceremony without
clarity.

## Tests

All 1,568 unit tests pass locally. Ruff lint + format clean.
